+++ src/main/java/legend/game/wmap/WMapModelAndAnimData258.java
@@ -106,7 +106,7 @@ public class WMapModelAndAnimData258 {
   public WMapAtmosphericEffectInstance60[] atmosphericEffectInstances_24;
   public MeshObj[] atmosphericEffectSprites;
 
-  public int clutYIndex_28;
+  public float clutYIndex_28;
   public FileData imageData_2c;
   public FileData imageData_30;
   /** Used as the camera position, only translation is used */
